test: move createConflict() from fixtures to util
The files in test/integration/fixtures are repeatedly `require()`d and run
in the `initialize()` function of `test/integration/setup.js`.  The
`scenarios.js` does not need this treatment.
